backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
vaderSentiment==3.3.2
scikit-learn>=1.5.0
numpy>=2.0.0
requests==2.31.0
beautifulsoup4==4.12.3
playwright>=1.48.0
spacy>=3.7.4
python-dotenv==1.0.1
httpx==0.27.0
lxml>=5.3.0
